Renaming project to have MindTouch prefix
