+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer.Test/Helpers/CodeFixVerifier.Helper.cs
