+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer.Test/Helpers/DiagnosticResult.cs
