+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer.Test/Helpers/DiagnosticVerifier.Helper.cs
