+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer.Test/Verifiers/DiagnosticVerifier.cs
