+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer.Test/packages.config
