+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer.Vsix/MindTouchEnumSwitchAnalyzer.Vsix.csproj
@@ -13,8 +13,8 @@
     <ProjectGuid>{11BB92A0-7C56-44E6-BE63-55AE8F0185D5}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>EnumSwitchAnalyzer.Vsix</RootNamespace>
-    <AssemblyName>EnumSwitchAnalyzer.Vsix</AssemblyName>
+    <RootNamespace>MindTouchEnumSwitchAnalyzer.Vsix</RootNamespace>
+    <AssemblyName>MindTouchEnumSwitchAnalyzer.Vsix</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <GeneratePkgDefFile>false</GeneratePkgDefFile>
     <IncludeAssemblyInVSIXContainer>false</IncludeAssemblyInVSIXContainer>
@@ -52,9 +52,11 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\EnumSwitchAnalyzer\EnumSwitchAnalyzer.csproj">
+    <ProjectReference Include="..\MindTouchEnumSwitchAnalyzer\MindTouchEnumSwitchAnalyzer.csproj">
       <Project>{1FC79BE5-ABF5-475C-A62C-2BD7BE9B6EE8}</Project>
-      <Name>EnumSwitchAnalyzer</Name>
+      <Name>MindTouchEnumSwitchAnalyzer</Name>
+      <IncludeOutputGroupsInVSIX>BuiltProjectOutputGroup%3bBuiltProjectOutputGroupDependencies%3bGetCopyToOutputDirectoryItems%3bSatelliteDllsProjectOutputGroup%3b</IncludeOutputGroupsInVSIX>
+      <IncludeOutputGroupsInVSIXLocalOnly>DebugSymbolsProjectOutputGroup%3b</IncludeOutputGroupsInVSIXLocalOnly>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
