+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer.Vsix/source.extension.vsixmanifest
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <PackageManifest Version="2.0.0" xmlns="http://schemas.microsoft.com/developer/vsx-schema/2011" xmlns:d="http://schemas.microsoft.com/developer/vsx-schema-design/2011">
   <Metadata>
-    <Identity Id="EnumSwitchAnalyzer..941c8d45-2933-427c-9cd8-96d94a7cfac2" Version="1.0" Language="en-US" Publisher="Admin"/>
-    <DisplayName>EnumSwitchAnalyzer</DisplayName>
+    <Identity Id="MindTouchEnumSwitchAnalyzer..941c8d45-2933-427c-9cd8-96d94a7cfac2" Version="1.0" Language="en-US" Publisher="Admin"/>
+    <DisplayName>MindTouchEnumSwitchAnalyzer</DisplayName>
     <Description xml:space="preserve">Roslyn analyzer that ensures that all enum cases are handled when used inside a switch statement.</Description>
   </Metadata>
   <Installation>
@@ -12,7 +12,7 @@
     <Dependency Id="Microsoft.Framework.NDP" DisplayName="Microsoft .NET Framework" d:Source="Manual" Version="[4.5,)" />
   </Dependencies>
   <Assets>
-    <Asset Type="Microsoft.VisualStudio.MefComponent" d:Source="Project" d:ProjectName="EnumSwitchAnalyzer" Path="|EnumSwitchAnalyzer|"/>
-    <Asset Type="Microsoft.VisualStudio.Analyzer" d:Source="Project" d:ProjectName="EnumSwitchAnalyzer" Path="|EnumSwitchAnalyzer|"/>
+    <Asset Type="Microsoft.VisualStudio.MefComponent" d:Source="Project" d:ProjectName="MindTouchEnumSwitchAnalyzer" Path="|MindTouchEnumSwitchAnalyzer|" />
+    <Asset Type="Microsoft.VisualStudio.Analyzer" d:Source="Project" d:ProjectName="MindTouchEnumSwitchAnalyzer" Path="|MindTouchEnumSwitchAnalyzer|" />
   </Assets>
 </PackageManifest>
