+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer/CodeFixProvider.cs
@@ -22,12 +22,12 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CodeActions;
+using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace EnumSwitchAnalyzer {
+namespace MindTouchEnumSwitchAnalyzer {
     [ExportCodeFixProvider(LanguageNames.CSharp, Name = nameof(EnumSwitchAnalyzerCodeFixProvider)), Shared]
     public class EnumSwitchAnalyzerCodeFixProvider : CodeFixProvider {
