+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer/Diagnostic.nuspec
@@ -1,19 +1,19 @@
 ﻿<?xml version="1.0"?>
 <package xmlns="http://schemas.microsoft.com/packaging/2011/08/nuspec.xsd">
   <metadata>
-    <id>EnumSwitchAnalyzer</id>
+    <id>MindTouchEnumSwitchAnalyzer</id>
     <version>1.0.0.0</version>
-    <title>EnumSwitchAnalyzer</title>
+    <title>MindTouchEnumSwitchAnalyzer</title>
     <authors>Admin</authors>
     <owners>Admin</owners>
     <licenseUrl>https://github.com/MindTouch/RoslynAnalyzers/blob/master/LICENSE</licenseUrl>
     <projectUrl>https://github.com/MindTouch/RoslynAnalyzers/</projectUrl>
     <iconUrl>http://ICON_URL_HERE_OR_DELETE_THIS_LINE</iconUrl>
     <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>EnumSwitchAnalyzer</description>
+    <description>MindTouchEnumSwitchAnalyzer</description>
     <releaseNotes>Summary of changes made in this release of the package.</releaseNotes>
     <copyright>Copyright</copyright>
-    <tags>EnumSwitchAnalyzer, analyzers</tags>
+    <tags>MindTouchEnumSwitchAnalyzer, analyzers</tags>
     <frameworkAssemblies>
       <frameworkAssembly assemblyName="System" targetFramework="" />
     </frameworkAssemblies>
