+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer/EnumSwitchAnalysis.cs
@@ -21,7 +21,7 @@ using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace EnumSwitchAnalyzer {
+namespace MindTouchEnumSwitchAnalyzer {
     internal class EnumSwitchAnalysis {
 
         //--- Class Methods ---
