+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer/Properties/AssemblyInfo.cs
@@ -4,11 +4,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("EnumSwitchAnalyzer")]
+[assembly: AssemblyTitle("MindTouchEnumSwitchAnalyzer")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("EnumSwitchAnalyzer")]
+[assembly: AssemblyProduct("MindTouchEnumSwitchAnalyzer")]
 [assembly: AssemblyCopyright("Copyright ©  2016")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
