+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer/Resources.resx
