+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer/packages.config
