+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer/tools/install.ps1
