+++ MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer/tools/uninstall.ps1
