+++ README.md
@@ -1,28 +1,28 @@
 RoslynAnalyzers
 ===============
+
 C# code analyzers built on the Roslyn compiler.
 
-EnumSwitchAnalyzer
-------------------
-Roslyn analyzer that ensures that all enum cases are handled when used inside a switch statement. [Read more.](EnumSwitchAnalyzer/EnumSwitchAnalyzer/README.md)
+MindTouchEnumSwitchAnalyzer
+---------------------------
+
+Roslyn analyzer that ensures that all enum cases are handled when used inside a switch statement. [Read more.](MindTouchEnumSwitchAnalyzer/MindTouchEnumSwitchAnalyzer/README.md)
 
 System Requirements
 -------------------
-* To run
-	* Microsoft Visual Studio 2015 with .Net 4.0 or later
-* To build
-	* Microsoft Visual Studio 2015 with .Net 4.5 or later
+
+-	To run
+	-	Microsoft Visual Studio 2015 with .Net 4.0 or later
+-	To build
+	-	Microsoft Visual Studio 2015 with .Net 4.5 or later
 
 License
 -------
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
 
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+```
+http://www.apache.org/licenses/LICENSE-2.0
+```
+
+Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
